pyproject.toml
[tool.poetry]
name = "musicplatform"
version = "0.1.0"
description = ""
authors = ["AbdulrhmanSayedAli <roctes7@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.1.2"
django-environ = "^0.9.0"
django-model-utils = "^4.2.0"
djangorestframework = "^3.14.0"
Pillow = "^9.2.0"
django-imagekit = "^4.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
